+++ src/command_modules/azure-cli-sql/azure/cli/command_modules/sql/_util.py
@@ -24,33 +24,51 @@ def get_sql_management_client(_):
     return get_mgmt_service_client(SqlManagementClient)
 
 
-def get_sql_servers_operations(kwargs):
-    return get_sql_management_client(kwargs).servers
+def get_sql_server_azure_ad_administrators_operations(kwargs):
+    return get_sql_management_client(kwargs).server_azure_ad_administrators
 
 
 def get_sql_capabilities_operations(kwargs):
     return get_sql_management_client(kwargs).capabilities
 
 
-def get_sql_firewall_rules_operations(kwargs):
-    return get_sql_management_client(kwargs).firewall_rules
-
-
 def get_sql_databases_operations(kwargs):
     return get_sql_management_client(kwargs).databases
+
+
+def get_sql_database_blob_auditing_policies_operations(kwargs):
+    return get_sql_management_client(kwargs).database_blob_auditing_policies
+
+
+def get_sql_database_threat_detection_policies_operations(kwargs):
+    return get_sql_management_client(kwargs).database_threat_detection_policies
+
+
+def get_sql_database_usages_operations(kwargs):
+    return get_sql_management_client(kwargs).database_usages
 
 
 def get_sql_elastic_pools_operations(kwargs):
     return get_sql_management_client(kwargs).elastic_pools
 
 
+def get_sql_firewall_rules_operations(kwargs):
+    return get_sql_management_client(kwargs).firewall_rules
+
+
 def get_sql_recommended_elastic_pools_operations(kwargs):
     return get_sql_management_client(kwargs).recommended_elastic_pools
 
+def get_sql_replication_links_operations(kwargs):
+    return get_sql_management_client(kwargs).replication_links
 
-def get_sql_server_azure_ad_administrators_operations(kwargs):
-    return get_sql_management_client(kwargs).server_azure_ad_administrators
+
+def get_sql_servers_operations(kwargs):
+    return get_sql_management_client(kwargs).servers
 
 
+def get_sql_server_usages_operations(kwargs):
+    return get_sql_management_client(kwargs).server_usages
+  
 def get_sql_virtual_network_rules_operations(kwargs):
     return get_sql_management_client(kwargs).virtual_network_rules