+++ src/command_modules/azure-cli-sql/azure/cli/command_modules/sql/commands.py
@@ -8,13 +8,19 @@ from azure.cli.core.sdk.util import (
     create_service_adapter,
     ServiceGroup)
 from ._util import (
-    get_sql_servers_operations,
-    get_sql_firewall_rules_operations,
-    get_sql_databases_operations,
-    get_sql_elastic_pools_operations,
     get_sql_server_azure_ad_administrators_operations,
     get_sql_capabilities_operations,
-    get_sql_virtual_network_rules_operations)
+    get_sql_databases_operations,
+    get_sql_database_blob_auditing_policies_operations,
+    get_sql_database_threat_detection_policies_operations,
+    get_sql_database_usages_operations,
+    get_sql_elastic_pools_operations,
+    get_sql_firewall_rules_operations,
+    get_sql_replication_links_operations,
+    get_sql_servers_operations,
+    get_sql_server_usages_operations,
+    get_sql_virtual_network_rules_operations
+)
 
 if not supported_api_version(PROFILE_TYPE, max_api='2017-03-09-profile'):
     custom_path = 'azure.cli.command_modules.sql.custom#{}'
@@ -56,9 +62,6 @@ if not supported_api_version(PROFILE_TYPE, max_api='2017-03-09-profile'):
 
         with s.group('sql db replica') as c:
             c.custom_command('create', 'db_create_replica')
-            c.command('list-links', 'list_replication_links')
-            c.custom_command('delete-link', 'db_delete_replica_link', confirmation=True)
-            c.custom_command('set-primary', 'db_failover')
 
         with s.group('sql dw') as c:
             c.custom_command('create', 'dw_create')
@@ -86,17 +89,43 @@ if not supported_api_version(PROFILE_TYPE, max_api='2017-03-09-profile'):
         #     c.command('list', 'list_service_tier_advisors')
         #     c.command('show', 'get_service_tier_advisor')
 
+    replication_links_operations = create_service_adapter('azure.mgmt.sql.operations.replication_links_operations',
+                                                          'ReplicationLinksOperations')
+    with ServiceGroup(__name__, get_sql_replication_links_operations,
+                      replication_links_operations, custom_path) as s:
+        with s.group('sql db replica') as c:
+            c.command('list-links', 'list_by_database')
+            c.custom_command('delete-link', 'db_delete_replica_link', confirmation=True)
+            c.custom_command('set-primary', 'db_failover')
+
+    database_blob_auditing_policies_operations = create_service_adapter(
+        'azure.mgmt.sql.operations.database_blob_auditing_policies_operations',
+        'DatabaseBlobAuditingPoliciesOperations')
+    with ServiceGroup(__name__, get_sql_database_blob_auditing_policies_operations,
+                      database_blob_auditing_policies_operations, custom_path) as s:
         with s.group('sql db audit-policy') as c:
-            c.command('show', 'get_blob_auditing_policy')
+            c.command('show', 'get')
             c.generic_update_command(
-                'update', 'get_blob_auditing_policy', 'create_or_update_blob_auditing_policy',
+                'update', 'get', 'create_or_update',
                 custom_func_name='db_audit_policy_update')
 
+    database_threat_detection_policies_operations = create_service_adapter(
+        'azure.mgmt.sql.operations.database_threat_detection_policies_operations',
+        'DatabaseThreatDetectionPoliciesOperations')
+    with ServiceGroup(__name__, get_sql_database_threat_detection_policies_operations,
+                      database_threat_detection_policies_operations, custom_path) as s:
         with s.group('sql db threat-policy') as c:
-            c.command('show', 'get_threat_detection_policy')
-            c.generic_update_command('update', 'get_threat_detection_policy',
-                                     'create_or_update_threat_detection_policy',
+            c.command('show', 'get')
+            c.generic_update_command('update', 'get',
+                                     'create_or_update',
                                      custom_func_name='db_threat_detection_policy_update')
+
+    database_usages_operations = create_service_adapter('azure.mgmt.sql.operations.database_usages_operations',
+                                                        'DatabaseUsagesOperations')
+
+    with ServiceGroup(__name__, get_sql_database_usages_operations, database_usages_operations, custom_path) as s:
+        with s.group('sql db') as c:
+            c.command('list-usages', 'list_by_database')
 
     ###############################################
     #                sql elastic-pool             #
@@ -111,10 +140,13 @@ if not supported_api_version(PROFILE_TYPE, max_api='2017-03-09-profile'):
             c.command('delete', 'delete')
             c.command('show', 'get')
             c.command('list', 'list_by_server')
-            c.command('list-dbs', 'list_databases')
             c.generic_update_command(
                 'update', 'get', 'create_or_update',
                 custom_func_name='elastic_pool_update')
+
+    with ServiceGroup(__name__, get_sql_databases_operations, database_operations, custom_path) as s:
+        with s.group('sql elastic-pool') as c:
+            c.command('list-dbs', 'list_by_elastic_pool')
 
     recommanded_elastic_pools_ops = \
         create_service_adapter('azure.mgmt.sql.operations.recommended_elastic_pools_operations',
@@ -144,10 +176,16 @@ if not supported_api_version(PROFILE_TYPE, max_api='2017-03-09-profile'):
             c.command('create', 'create_or_update')
             c.command('delete', 'delete', confirmation=True)
             c.command('show', 'get')
-            c.command('list-usages', 'list_usages')
             c.command('list', 'list_by_resource_group')
             c.generic_update_command('update', 'get', 'create_or_update',
                                      custom_func_name='server_update')
+
+    server_usages_operations = create_service_adapter('azure.mgmt.sql.operations.server_usages_operations',
+                                                      'ServerUsagesOperations')
+
+    with ServiceGroup(__name__, get_sql_server_usages_operations, server_usages_operations, custom_path) as s:
+        with s.group('sql server') as c:
+            c.command('list-usages', 'list_by_server')
 
     firewall_rules_operations = create_service_adapter(
         'azure.mgmt.sql.operations.firewall_rules_operations',
@@ -172,7 +210,7 @@ if not supported_api_version(PROFILE_TYPE, max_api='2017-03-09-profile'):
                       aadadmin_operations, custom_path) as s:
         with s.group('sql server ad-admin') as c:
             c.custom_command('create', 'server_ad_admin_set')
-            c.command('list', 'list')
+            c.command('list', 'list_by_server')
             c.command('delete', 'delete')
             c.generic_update_command('update', 'get', 'create_or_update',
                                      custom_func_name='server_ad_admin_update',