+++ src/command_modules/azure-cli-acs/HISTORY.rst
@@ -2,6 +2,11 @@
 
 Release History
 ===============
+unreleased
++++++++++++++++++++
+* add acs list-locations command
+* make ssh-key-file come with expected default value
+
 2.0.14 (2017-08-28)
 +++++++++++++++++++
 * correct preview regions