+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -2,6 +2,14 @@
 
 Release History
 ===============
+(unreleased)
++++++++++++++++++++
+* webapp: able to create a webapp in a resource group other than the service plan's
+
+0.1.15 (2017-08-31)
++++++++++++++++++++
+* minor fixes
+
 0.1.14 (2017-08-28)
 +++++++++++++++++++
 Breaking Change:webapp: fix inconsistencies in the output of "az webapp config appsettings delete/set"
@@ -9,6 +17,8 @@ webapp: add a new alias of '-i' for "az webapp config container set --docker-cus
 webapp: expose 'az webapp log show'
 webapp: expose new arguments from 'az webapp delete' to retain app service plan, metrics or dns registration. 
 webapp: detect a slot setting correctly 
+webapp: add param --docker-container-logging that goes through the same logic as --web-server-logging
+webapp: add premium v2 sku
 
 0.1.13 (2017-08-15)
 +++++++++++++++++++