+++ src/command_modules/azure-cli-sf/HISTORY.rst
@@ -2,6 +2,11 @@
 
 Release History
 ===============
+
+1.0.8 (2017-08-31)
+++++++++++++++++++
+* Deprecating all commands in favor of Service Fabric CLI (sfctl)
+
 1.0.7 (2017-08-28)
 ++++++++++++++++++
 * Simplified registry user/pass rules for command.