+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -2,6 +2,13 @@
 
 Release History
 ===============
+(unreleased)
++++++++++++++++++++
+* msi: don't assign access unless `--scope` is provided
+* msi: use the same extension naming as portal does
+* msi: remove the useless `subscription` from the `vm/vmss create` commands output
+* `vm/vmss create`: fix a bug that the storage sku is not applied on data disks coming with an image
+
 2.0.13 (2017-08-28)
 +++++++++++++++++++
 * `vmss get-instance-view`: Fix issue where extra, erroneous information was displayed when using `--instance-id *`